backend/requirements.txt
django>=5.2,<5.3
djangorestframework>=3.17,<3.18
djangorestframework-simplejwt>=5.5,<6
django-cors-headers>=4.9,<5
PyMySQL>=1.1,<2
redis>=7.4,<8
celery>=5.6,<6
gunicorn>=23,<24
